fixes: "Correccion en services y controller"

+++ src/main/java/com/crud/crudprueba/data/crud/ICarCrudRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICarCrudRepository extends CrudRepository<Car, Integer> {
+public interface ICarCrudRepository extends CrudRepository<Car, Long> {
 }
