+++ src/main/java/com/crud/crudprueba/data/crud/IPaymentCrudRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPaymentCrudRepository extends CrudRepository<Payment, Integer> {
+public interface IPaymentCrudRepository extends CrudRepository<Payment, Long> {
